+++ apps/web/src/components/SlotPicker/SlotPicker.tsx
@@ -1,0 +1,87 @@
+import { Slot } from '@internship-app/types';
+import { Box, useMediaQuery } from '@mui/material';
+import { useState } from 'react';
+
+import { ConfirmDialog } from '../ConfirmDialog';
+import { DatePicker, MuiDate } from './DatePicker';
+import { Layout } from './Layout';
+import { TimeSlotPicker } from './TimeSlotPicker';
+
+type SlotPickerProps = {
+  title: string;
+  slots: Slot[];
+  handleSubmit: (slot: Slot) => void;
+};
+
+const SlotPicker: React.FC<SlotPickerProps> = ({
+  title,
+  slots,
+  handleSubmit,
+}) => {
+  const [selectedDate, setSelectedDate] = useState<MuiDate | null>(null);
+  const [selectedSlot, setSelectedSlot] = useState<Slot | null>(null);
+  const [dialogOpen, setDialogOpen] = useState(false);
+
+  const isMobile = useMediaQuery('(max-width:700px)');
+
+  const testSlotDateFormat =
+    selectedSlot?.start.toLocaleString('hr-HR', {
+      timeStyle: 'short',
+      dateStyle: 'short',
+    }) +
+    '-' +
+    selectedSlot?.end.toLocaleTimeString('hr-HR', {
+      timeStyle: 'short',
+    });
+
+  if (slots?.length === 0) {
+    return (
+      <Layout title="Zasad nema dostupnih termina, ali obavijestit ćemo te mailom kad ih bude." />
+    );
+  }
+
+  return (
+    <Layout title={title}>
+      <Box
+        sx={{
+          display: 'flex',
+          alignItems: isMobile ? 'center' : 'flex-start',
+          flexDirection: isMobile ? 'column' : 'row',
+        }}
+      >
+        <DatePicker
+          availableDates={slots?.map((s) => s.start) || []}
+          onChange={setSelectedDate}
+        />
+        {!!selectedDate && (
+          <TimeSlotPicker
+            availableTimeSlots={
+              slots?.filter(
+                (slot) =>
+                  slot.start.getDate() === selectedDate.$D &&
+                  slot.start.getMonth() === selectedDate.$M &&
+                  slot.start.getFullYear() === selectedDate.$y,
+              ) || []
+            }
+            isMobile={isMobile}
+            onChange={(slot) => {
+              setSelectedSlot(slot);
+              setDialogOpen(true);
+            }}
+          />
+        )}
+      </Box>
+      <ConfirmDialog
+        open={!!dialogOpen}
+        handleClose={(confirmed) => {
+          if (confirmed) handleSubmit(selectedSlot!);
+          setDialogOpen(false);
+        }}
+        title="Potvrdi odabir termina"
+        description={`Vaš termin bit će rezerviran za ${testSlotDateFormat}.`}
+      />
+    </Layout>
+  );
+};
+
+export default SlotPicker;
