+++ apps/web/src/components/SlotPicker/index.ts
@@ -1,0 +1,5 @@
+import { Layout } from './Layout';
+import SlotPicker from './SlotPicker';
+
+export { Layout as SlotPickerLayout };
+export default SlotPicker;
