--- apps/web/src/components/SlotPickerLayout/index.ts
@@ -1,3 +1,0 @@
-import SlotPickerLayout from './SlotPickerLayout';
-
-export default SlotPickerLayout;
