--- apps/web/src/components/SlotPickers/index.ts
@@ -1,2 +1,0 @@
-export * from './DatePicker';
-export * from './TimeSlotPicker';
