+++ apps/web/src/pages/ScheduleTestPage/ScheduleTestPage.tsx
@@ -1,17 +1,9 @@
 import { Discipline, Slot } from '@internship-app/types';
-import { Box, useMediaQuery } from '@mui/material';
-import { useState } from 'react';
 import { useRoute } from 'wouter';
 
 import { useFetchAvailableTestSlots } from '../../api/useFetchAvailableTestSlots';
 import { useScheduleTest } from '../../api/useScheduleTest';
-import { ConfirmDialog } from '../../components/ConfirmDialog';
-import SlotPickerLayout from '../../components/SlotPickerLayout';
-import {
-  DatePicker,
-  MuiDate,
-  TimeSlotPicker,
-} from '../../components/SlotPickers';
+import SlotPicker, { SlotPickerLayout } from '../../components/SlotPicker';
 import { disciplineLabel } from '../../constants/internConstants';
 import { Path } from '../../constants/paths';
 
@@ -19,7 +11,6 @@ type Params = { internId: string; discipline: Discipline };
 
 const ScheduleTestPage = () => {
   const [, params] = useRoute<Params>(Path.ScheduleTest);
-  const isMobile = useMediaQuery('(max-width:700px)');
 
   const {
     data: slots,
@@ -30,11 +21,7 @@ const ScheduleTestPage = () => {
 
   const scheduleTest = useScheduleTest();
 
-  const [selectedDate, setSelectedDate] = useState<MuiDate | null>(null);
-  const [selectedSlot, setSelectedSlot] = useState<Slot | null>(null);
-  const [dialogOpen, setDialogOpen] = useState(false);
-
-  const handleSubmit = () => {
+  const handleSubmit = (selectedSlot: Slot) => {
     if (!selectedSlot || !params?.internId) return;
 
     scheduleTest.mutate({
@@ -43,76 +30,23 @@ const ScheduleTestPage = () => {
     });
   };
 
-  const testSlotDateFormat =
-    selectedSlot?.start.toLocaleString('hr-HR', {
-      timeStyle: 'short',
-      dateStyle: 'short',
-    }) +
-    '-' +
-    selectedSlot?.end.toLocaleTimeString('hr-HR', {
-      timeStyle: 'short',
-    });
-
   if (isLoading) return <SlotPickerLayout title="Loading..." />;
 
-  if (isError || !params?.discipline || !params.internId)
+  if (isError || !params?.discipline || !params.internId || !slots)
     return (
       <SlotPickerLayout
         title={`Dogodila se greška (${error}). Molimo kontaktirajte nas na info@dump.hr`}
       />
     );
 
-  if (slots?.length === 0) {
-    return (
-      <SlotPickerLayout title="Zasad nema dostupnih termina, ali obavijestit ćemo te mailom kad ih bude." />
-    );
-  }
-
   return (
-    <SlotPickerLayout
+    <SlotPicker
       title={`Odaberi termin za ispit iz područja ${
         disciplineLabel[params.discipline]
       }`}
-    >
-      <Box
-        sx={{
-          display: 'flex',
-          alignItems: isMobile ? 'center' : 'flex-start',
-          flexDirection: isMobile ? 'column' : 'row',
-        }}
-      >
-        <DatePicker
-          availableDates={slots?.map((s) => s.start) || []}
-          onChange={setSelectedDate}
-        />
-        {!!selectedDate && (
-          <TimeSlotPicker
-            availableTimeSlots={
-              slots?.filter(
-                (slot) =>
-                  slot.start.getDate() === selectedDate.$D &&
-                  slot.start.getMonth() === selectedDate.$M &&
-                  slot.start.getFullYear() === selectedDate.$y,
-              ) || []
-            }
-            isMobile={isMobile}
-            onChange={(slot) => {
-              setSelectedSlot(slot);
-              setDialogOpen(true);
-            }}
-          />
-        )}
-      </Box>
-      <ConfirmDialog
-        open={!!dialogOpen}
-        handleClose={(confirmed) => {
-          if (confirmed) handleSubmit();
-          setDialogOpen(false);
-        }}
-        title="Potvrdi odabir termina"
-        description={`Vaš termin bit će rezerviran za ${testSlotDateFormat}.`}
-      />
-    </SlotPickerLayout>
+      slots={slots}
+      handleSubmit={handleSubmit}
+    />
   );
 };
